Add icons and update JS events for pen tool
Add icons to HTML from Google Fonts to enhance UX. Style them with CSS.
Add event listeners & switch to pen tool to listen for mousedown &
mouseup to activate/terminate the tool. The pen tool is glitchy, and it
sometimes does not stop once the mouse is lifted. Debugging this will
be targeted next.