Add rainbow pen tool functionality
Create an event listener for the rainbow pen button that creates a
rainbow stroke when pressed. This was accomplished by creating a
function that generates a hex code based on random selection of indices
within an array of hex characters, then placing said function within a
variation of the plain pen tool button.