Fix stroke color bug in clear button event
When drawing after using the "clear" button, the mouse will sometimes
color a rainbow or white cell rather than black. Fix the clear button's
event listener so that it defaults to the previously used stroke tool.

Update the grid button's alert message so that the user knows it will
default to the black pen tool upon resize.